Moves widget configuration into a standalone widget config provider model.
