+++ view/frontend/templates/checkout.phtml
@@ -5,7 +5,7 @@
 
     <script>
         require(['magento_plugin', 'addressfinder', 'domReady!'], function (MagentoPlugin, AddressFinder) {
-          var widgetConfig = <?=json_encode($block->getConfig()) ?>;
+          var widgetConfig = <?=json_encode($block->getWidgetConfig()) ?>;
 
           window.AddressFinder = AddressFinder;
           new MagentoPlugin(widgetConfig);
