+++ view/frontend/templates/customer.phtml
@@ -6,7 +6,7 @@
 
     <script>
         require(['magento_plugin', 'addressfinder', 'domReady!'], function (MagentoPlugin, AddressFinder) {
-            var widgetConfig = <?=json_encode($block->getConfig()) ?>;
+            var widgetConfig = <?=json_encode($block->getWidgetConfig()) ?>;
 
             window.AddressFinder = AddressFinder;
             new MagentoPlugin(widgetConfig);
